workaround mismatch between API and Pagewise controller. (#23)
When filters are used, the API can return pages with fewer responses than the requested page size. The Pagewise controller interprets this as the end of the result set. This makes repeated calls for pages until the page size is obtained or there are no more results. Anything over the page size is "carried" forward to the next page request.

This change is only made for APIs, since these are by far the largest collections in our current datasets and the only ones where this is a problem. We may eliminate the need for this with changes to the Registry API.